优化代码,提高效率,去除大部分的GC问题

+++ Assets/Scripts/Frame/Character/CharacterBaseData.cs
@@ -5,5 +5,5 @@ using System.Collections.Generic;
 public class CharacterBaseData : GameBase
 {
 	public string mName;
-	public uint	mGUID;		// 玩家唯一ID,由服务器发送过来的
+	public ulong mGUID;		// 玩家唯一ID,由服务器发送过来的
 }
