+++ Assets/Scripts/Frame/Socket/SocketPacket.cs
@@ -4,18 +4,18 @@ using System.Collections.Generic;
 
 public class SocketPacket : SerializedData, IClassObject
 {
-	public int mType;
+	public ushort mType;
 	public ISocketConnect mConnect; // 如果是从服务器发过来的消息,记录了是从哪个服务器发送过来的
 	public NetClient mClient;       // 如果是从客户端发过来的消息,记录了是从哪个客户端发过来的
 	public uint mClientID;			// 客户端ID
 	public void setConnect(ISocketConnect connect) { mConnect = connect; }
-	public virtual void init(int type)
+	public virtual void init(ushort type)
 	{
 		mType = type;
 		fillParams();
 		zeroParams();
 	}
-	public int getPacketType() { return mType; }
+	public ushort getPacketType() { return mType; }
 	// 如果是服务器向客户端发送的消息,则需要重写该函数
 	public virtual void execute() { }
 	protected override void fillParams() { }
